Move runit service handling to libpiksi
