+++ package/libpiksi/libpiksi/src/Makefile
@@ -1,5 +1,13 @@
 TARGET=libpiksi
-SOURCES=sbp_zmq_tx.c sbp_zmq_rx.c sbp_zmq_pubsub.c settings.c util.c logging.c networking.c
+SOURCES=\
+  sbp_zmq_tx.c \
+  sbp_zmq_rx.c \
+  sbp_zmq_pubsub.c \
+  settings.c \
+  util.c \
+  logging.c \
+  networking.c \
+  runit.c
 CFLAGS=-std=gnu11 -fPIC -I../include
 ARFLAGS=rcs
 LDFLAGS=-shared
